Week10/Day2/requirements.txt
langgraph
langchain
langchain_community
langchain_groq
langchain_huggingface
streamlit
tavily-python
langchain-tavily
faiss-cpu
python-dotenv
sentence-transformers